+++ dragonite-forwarder/src/main/java/com/vecsight/dragonite/forwarder/misc/ForwarderGlobalConstants.java
@@ -26,7 +26,7 @@ public final class ForwarderGlobalConstants {
 
     //Protocol shits
 
-    public static final byte PROTOCOL_VERSION = 2;
+    public static final byte PROTOCOL_VERSION = 3;
 
     public static final int DEFAULT_SERVER_PORT = 5233;