+++ dragonite-proxy/src/main/java/com/vecsight/dragonite/proxy/config/ProxyClientConfig.java
@@ -16,6 +16,7 @@ package com.vecsight.dragonite.proxy.config;
 import com.vecsight.dragonite.proxy.acl.ParsedACL;
 import com.vecsight.dragonite.proxy.misc.ProxyGlobalConstants;
 import com.vecsight.dragonite.sdk.config.DragoniteSocketParameters;
+import com.vecsight.dragonite.sdk.obfs.Obfuscator;
 
 import java.net.InetSocketAddress;
 
@@ -131,5 +132,13 @@ public class ProxyClientConfig {
 
     public DragoniteSocketParameters getDragoniteSocketParameters() {
         return dragoniteSocketParameters;
+    }
+
+    public void setObfuscator(final Obfuscator obfuscator) {
+        dragoniteSocketParameters.setObfuscator(obfuscator);
+    }
+
+    public Obfuscator getObfuscator() {
+        return dragoniteSocketParameters.getObfuscator();
     }
 }