+++ dragonite-sdk/src/main/java/com/vecsight/dragonite/sdk/obfs/Obfuscator.java
@@ -1,0 +1,24 @@
+/*
+ * VECTORSIGHT CONFIDENTIAL
+ * ------------------------
+ * Copyright (c) [2015] - [2017]
+ * VectorSight Systems Co., Ltd.
+ * All Rights Reserved.
+ *
+ * Unauthorized copying of this file, via any medium is strictly prohibited
+ * Proprietary and confidential
+ *
+ * Written by Toby Huang <t@vecsight.com>, June 2017
+ */
+
+package com.vecsight.dragonite.sdk.obfs;
+
+public interface Obfuscator {
+
+    byte[] obfuscate(final byte[] rawData);
+
+    byte[] deobfuscate(final byte[] obfsData);
+
+    int getReceiveBufferOverhead();
+
+}