+++ dragonite-sdk/src/main/java/com/vecsight/dragonite/sdk/socket/DragoniteServer.java
@@ -21,6 +21,7 @@ import com.vecsight.dragonite.sdk.msg.MessageParser;
 import com.vecsight.dragonite.sdk.msg.ReliableMessage;
 import com.vecsight.dragonite.sdk.msg.types.ACKMessage;
 import com.vecsight.dragonite.sdk.msg.types.CloseMessage;
+import com.vecsight.dragonite.sdk.obfs.Obfuscator;
 import com.vecsight.dragonite.sdk.web.DevConsoleWebServer;
 
 import java.io.IOException;
@@ -42,6 +43,8 @@ public class DragoniteServer {
     private final int heartbeatIntervalSec, receiveTimeoutSec;
     private final boolean autoSplit;
     private final boolean enableWebPanel;
+    private final Obfuscator obfuscator;
+    private final int obfsOverhead;
     //end
 
     private volatile long defaultSendSpeed;
@@ -83,6 +86,8 @@ public class DragoniteServer {
         autoSplit = parameters.isAutoSplit();
         enableWebPanel = parameters.isEnableWebPanel();
         devConsoleBindAddress = parameters.getWebPanelBindAddress();
+        obfuscator = parameters.getObfuscator();
+        obfsOverhead = obfuscator != null ? obfuscator.getReceiveBufferOverhead() : 0;
         //end
 
         this.defaultSendSpeed = defaultSendSpeed;
@@ -96,7 +101,7 @@ public class DragoniteServer {
         receiveThread = new Thread(() -> {
             try {
                 while (doReceive) {
-                    final byte[] b = new byte[packetSize];
+                    final byte[] b = new byte[packetSize + obfsOverhead];
                     final DatagramPacket packet = new DatagramPacket(b, b.length);
                     try {
                         datagramSocket.receive(packet);
@@ -183,7 +188,8 @@ public class DragoniteServer {
         final SocketAddress remoteAddress = packet.getSocketAddress();
         Message message = null;
         try {
-            message = MessageParser.parseMessage(packet.getData());
+            final byte[] data = obfuscator != null ? obfuscator.deobfuscate(packet.getData()) : packet.getData();
+            if (data != null) message = MessageParser.parseMessage(data);
         } catch (final IncorrectMessageException ignored) {
         }
 
@@ -223,9 +229,12 @@ public class DragoniteServer {
 
     //SEND ALL PACKETS THROUGH THIS!!
     protected void sendPacket(final byte[] bytes, final SocketAddress socketAddress) throws IOException {
-        final DatagramPacket packet = new DatagramPacket(bytes, bytes.length);
-        packet.setSocketAddress(socketAddress);
-        datagramSocket.send(packet);
+        final byte[] data = obfuscator != null ? obfuscator.obfuscate(bytes) : bytes;
+        if (data != null) {
+            final DatagramPacket packet = new DatagramPacket(data, data.length);
+            packet.setSocketAddress(socketAddress);
+            datagramSocket.send(packet);
+        }
     }
 
     public long getDefaultSendSpeed() {